refactor(sandbox): localize runtime preparation
